+++ metadata.rb
@@ -5,11 +5,13 @@ license 'MIT'
 description 'Installs/Configures a development box for Chef/Kitchen/Cloud'
 long_description 'Installs/Configures Chefdk, VirtualBox, Packer, Vagrant,
     Atom editor, Habitat, and other key tools needed to get started hacking.'
-version '0.1.20'
+version '0.2.0'
 
 supports 'ubuntu'
 issues_url 'https://github.com/dayne/d_devbox/issues' if respond_to?(:issues_url)
 source_url 'https://github.com/dayne/d_devbox' if respond_to?(:source_url)
+
+depends 'apt'
 
 # chefdk related
 depends 'atom', '>= 0.2.0'
@@ -19,4 +21,4 @@ depends 'terraform', '>= 1.0.1'
 depends 'docker', '>= 2.15.25'
 
 # note: following have special branches to work - See Berksfile
-depends 'virtualbox', '>= 2.0'  # special branch
+depends 'virtualbox', '>= 2.1'  # special branch